feat(TextProcessing): add ability to defer text processing
The new TextProcessor concept allows the text processing to be deferred
to a Text Processor type. The Unity UI Text type is provided by default
as it is the current functionality. But it would now be possible and
trivial to make a Text Processor for TextMeshPro that could then be
used on the tooltip without needing to include TextMeshPro with this
repo.